Modified Box Size and Added Rule Merging
The original version had issues with creating boxes larger than 25 miles
on a side when you are far away from the equator. I fixed the command
line arguments that let you adjust that.

Also, Gnip has a limit on the number of rules you can provide. The
original version of bounding_boxes.py made a new rule for each 25x25
box, which quickly passes this limit if you are doing a big area (e.g.,
a country). Now, this script will merge 15 or so boxes into a single
rule using ORs to reduce the number of rules.